+++ Libraries/IO_Dynamic_Lib/ADMP_DecPreprocs_IodL_MOD.f08
@@ -1,5 +1,5 @@
 module ADMP_DecPreprocs
-    use, intrinsic :: iso_fortran_env, only : real32,real64,real128,int8
+    use, intrinsic :: iso_fortran_env, only : real32,real64,real128,int8,int16
 implicit none 
 !PREPROCESSING DIRECTIVES START FOR DECLARATION OF COORD AND VEL ARRAY 
 #ifdef PREC  
@@ -24,7 +24,7 @@ character(len=3) :: Units               ! UNITS IDENTIFIER FOR DISTANCE/ANGLES I
 
 ! G_LOG_INPUT SUB ACTUAL ARGUMENTS 
 integer(kind=int8) :: At_Num             ! NUMBER OF ATOMS
-integer(kind=int8) :: Step_Num           ! NUMBER OF STEPS
+integer(kind=int16) :: Step_Num           ! NUMBER OF STEPS
 logical :: IO_INPUT_ERROR=.FALSE.       ! FATAL ERROR FLAG FOR G_LOG_INPUT_SUBROUTINE 
 ! G_BIN_INPUT ACTUAL ARGUMENTS 
 character(len=132) :: Bin_filename
